+++ packages/plugin-packages/mashroom-portal-demo-alternative-theme/src/views/portal.tsx
@@ -6,7 +6,7 @@ import type {MashroomPortalPageRenderModel} from '@mashroom/mashroom-portal/type
 
 export default ({
                     user, site, siteBasePath, page, csrfToken, resourcesBasePath, apiBasePath,
-                    portalResourcesHeader, portalResourcesFooter, portalLayout, messages
+                    portalResourcesHeader, portalResourcesFooter, portalLayout, messages, lastThemeReloadTs
                 }: MashroomPortalPageRenderModel) => (
     <html>
         <head dangerouslySetInnerHTML={{ __html: `
@@ -23,10 +23,10 @@ export default ({
 
             <link href='https://fonts.googleapis.com/css?family=Roboto' rel='stylesheet' type='text/css'/>
             <link href='https://fonts.googleapis.com/css?family=Domine' rel='stylesheet' type='text/css'/>
-            <link rel="stylesheet" type="text/css" href='${resourcesBasePath}/bootstrap/css/bootstrap.css'/>
-            <link rel="stylesheet" type="text/css" href='${resourcesBasePath}/fontawesome/css/regular.css'/>
-            <link rel="stylesheet" type="text/css" href='${resourcesBasePath}/fontawesome/css/solid.css'/>
-            <link rel="stylesheet" type="text/css" href='${resourcesBasePath}/style.css'/>
+            <link rel="stylesheet" type="text/css" href='${resourcesBasePath}/bootstrap/css/bootstrap.css?v=${lastThemeReloadTs}'/>
+            <link rel="stylesheet" type="text/css" href='${resourcesBasePath}/fontawesome/css/regular.css?v=${lastThemeReloadTs}'/>
+            <link rel="stylesheet" type="text/css" href='${resourcesBasePath}/fontawesome/css/solid.css?v=${lastThemeReloadTs}'/>
+            <link rel="stylesheet" type="text/css" href='${resourcesBasePath}/style.css?v=${lastThemeReloadTs}'/>
 
             ${page.extraCss ? `<style type="text/css">${page.extraCss}</style>` : ''}